requirements.txt
# requirements.txt

# Airflow
apache-airflow==2.3.0
apache-airflow-providers-email
apache-airflow-providers-slack

# Machine Learning Libraries
transformers==4.21.0
datasets==2.3.2
torch==1.11.0
scikit-learn==0.24.2
pandas==1.3.0

# Other Dependencies
numpy==1.21.0

# Slack SDK (if using SlackOperator)
slack-sdk==3.9.1

# Additional Dependencies
python-dotenv
